Add a view mixin for audit resolve views
Use the new mixin in the existing views.
